+++ packages/antd-management-fast-common/src/components/TranslateLocale/index.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+import { formatMessage } from '../../utils/meta';
+
+export function TranslateLocale({ id, style }) {
+  return <div style={style}>{formatMessage({ id })}</div>;
+}