+++ packages/antd-management-fast-common/src/components/index.jsx
@@ -1,1 +1,2 @@
 export * from './InteractionAssemble';
+export * from './TranslateLocale';