+++ packages/antd-management-fast-framework/src/modelBuilders/index.js
@@ -6,6 +6,7 @@ import { buildModel as buildMenuControlModel } from './menuControl';
 import { buildModel as buildProgressBarControlModel } from './progressBarControl';
 import { buildModel as buildReloadAnimalPromptControlModel } from './reloadAnimalPromptControl';
 import { buildModel as buildSchedulingControlModel } from './schedulingControl';
+import { buildModel as buildShortcutControlModel } from './shortcutControl';
 import { buildModel as buildSwitchControlModel } from './switchControl';
 import { buildModel as buildTabControlModel } from './tabControl';
 
@@ -15,6 +16,7 @@ export function appendEmbedModelBuilder() {
   appendEmbedBuilder(buildProgressBarControlModel);
   appendEmbedBuilder(buildReloadAnimalPromptControlModel);
   appendEmbedBuilder(buildSchedulingControlModel);
+  appendEmbedBuilder(buildShortcutControlModel);
   appendEmbedBuilder(buildSwitchControlModel);
   appendEmbedBuilder(buildTabControlModel);
   appendEmbedBuilder(buildMenuControlModel);